+++ crates/ffi/CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## [Unreleased]
+
+### Added
+- `should_ignore_expires_for_local_urls` field to `Configuration` which allows to ignore expires for local urls during update. Default value is `false`.
+
 ## [2.0.1] - 2025-08-25
 
 ### Changed